Migrated to AspNetCore 2.1 preview
