+++ Erazer.Web.DocumentStore/Properties/launchSettings.json
@@ -4,7 +4,7 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:5002/",
-      "sslPort": 0
+      "sslPort": 50002
     }
   },
   "profiles": {
