+++ Erazer.Web.ReadAPI/Program.cs
@@ -8,16 +8,15 @@ namespace Erazer.Web.ReadAPI
     {
         public static void Main(string[] args)
         {
-            var host = BuildWebHost(args);
+            var host = CreateWebHostBuilder(args).Build();
             host.Services.Seed();
             host.Run();
         }
 
-        public static IWebHost BuildWebHost(string[] args) =>
+        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
                 .UseStartup<Startup>()
                 .UseApplicationInsights()
-                .UseUrls("http://localhost:5000")
-                .Build();
+                .UseUrls("http://localhost:5000");
     }
 }
